+++ src/interfaces/gridOrderableCell/gridOrderableCell.interface.ts
@@ -1,0 +1,12 @@
+import type {OrderableSADirective} from '../../directives/orderable/orderable.directive';
+
+/**
+ * Class used for storing orderable directive instance
+ */
+export interface GridOrderableCell
+{
+    /**
+     * Instance of orderable directive
+     */
+    orderable?: OrderableSADirective;
+}