+++ src/modules/matrixGrid.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 
 import {ContentContainerSAComponent, ContentRowContainerSAComponent, FooterContainerSAComponent, FooterRowContainerSAComponent, GridContainerSAComponent, HeaderContainerSAComponent, HeaderRowContainerSAComponent, MatrixGridSAComponent} from '../components';
-import {ContentCellTemplateSADirective, ContentContainerTemplateSADirective, ContentRowContainerTemplateSADirective, FooterCellTemplateSADirective, FooterContainerTemplateSADirective, FooterRowContainerTemplateSADirective, GridContainerTemplateSADirective, HeaderCellTemplateSADirective, HeaderContainerTemplateSADirective, HeaderRowContainerTemplateSADirective, MatrixGridColumnSADirective} from '../directives';
+import {ContentCellTemplateSADirective, ContentContainerTemplateSADirective, ContentRowContainerTemplateSADirective, FooterCellTemplateSADirective, FooterContainerTemplateSADirective, FooterRowContainerTemplateSADirective, GridContainerTemplateSADirective, HeaderCellTemplateSADirective, HeaderContainerTemplateSADirective, HeaderRowContainerTemplateSADirective, MatrixGridColumnSADirective, OrderableSADirective} from '../directives';
 
 /**
  * Module for matrix grid components and directives
@@ -29,6 +29,7 @@ import {ContentCellTemplateSADirective, ContentContainerTemplateSADirective, Con
         HeaderContainerTemplateSADirective,
         HeaderRowContainerTemplateSADirective,
         MatrixGridColumnSADirective,
+        OrderableSADirective,
     ],
     exports:
     [
@@ -51,6 +52,7 @@ import {ContentCellTemplateSADirective, ContentContainerTemplateSADirective, Con
         HeaderContainerTemplateSADirective,
         HeaderRowContainerTemplateSADirective,
         MatrixGridColumnSADirective,
+        OrderableSADirective,
     ],
 })
 export class MatrixGridModule